updatedb.cpp: Fix prunepaths getting missed.
If prune paths is
   /mnt/backup-1
   /mnt/backup-2

and the filesystem is:
   /bin/1
   /bin/2
   /mnt/backup-1/a
   /mnt/backup-1/b
   /mnt/backup-2/a
   /mnt/backup-2/b
   /run
   /var/lib
   /var/log

then conf_prunepaths_index gets ratcheted to the end of the prunepaths list
in the first scan() call, before it has recursed. Then when it _does_
recurse, the /mnt/backup-1 and /mnt/backup-2 prunepaths are never matched
because it has already moved past them.

This patch moves the conf_prunepaths_index to a local variable in scan() so
that it starts the list over every time it recurses.